refactor: share the temp dir helper
